addres code review
