+++ Test.Microsoft.NET.Build.Containers.Filesystem/CurrentFile.cs
@@ -7,6 +7,6 @@ public static class CurrentFile
     public static string Path([CallerFilePath] string file = "") => file;
 
     public static string Relative(string relative, [CallerFilePath] string file = "") {
-        return global::System.IO.Path.Combine(global::System.IO.Path.GetDirectoryName(file), relative);
+        return global::System.IO.Path.Combine(global::System.IO.Path.GetDirectoryName(file)!, relative); // file known to be not-null due to the mechanics of CallerFilePath
     }
 }
