+++ Test.Microsoft.NET.Build.Containers.Filesystem/ProjectInitializer.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using Microsoft.Build.Evaluation;
 using Microsoft.Build.Framework;
@@ -6,7 +5,7 @@ using Microsoft.Build.Locator;
 
 namespace Test.Microsoft.NET.Build.Containers;
 
-public static class Evaluator {
+public static class ProjectInitializer {
     private static string CombinedTargetsLocation;
 
     private static string CombineFiles(string propsFile, string targetsFile)
@@ -65,32 +64,5 @@ public static class Evaluator {
             props[kvp.Key] = kvp.Value;
         }
         return (collection.LoadProject(CombinedTargetsLocation, props, null), logs);
-    }
-}
-
-public class CapturingLogger : ILogger
-{
-    public LoggerVerbosity Verbosity { get => LoggerVerbosity.Diagnostic; set { } }
-    public string Parameters { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
-    private List<BuildMessageEventArgs> _messages = new();
-    public IReadOnlyList<BuildMessageEventArgs> Messages {get  { return _messages; } }
-
-    private List<BuildWarningEventArgs> _warnings = new();
-    public IReadOnlyList<BuildWarningEventArgs> Warnings {get  { return _warnings; } }
-
-    private List<BuildErrorEventArgs> _errors = new();
-    public IReadOnlyList<BuildErrorEventArgs> Errors {get  { return _errors; } }
-
-    public void Initialize(IEventSource eventSource)
-    {
-        eventSource.MessageRaised += (o, e) => _messages.Add(e);
-        eventSource.WarningRaised += (o, e) => _warnings.Add(e);
-        eventSource.ErrorRaised += (o, e) => _errors.Add(e);
-    }
-
-
-    public void Shutdown()
-    {
     }
 }
