+++ Test.Microsoft.NET.Build.Containers.Filesystem/Test.Microsoft.NET.Build.Containers.Filesystem.csproj
@@ -4,7 +4,6 @@
     <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <IsPackable>false</IsPackable>
-    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
