+++ Test.Microsoft.NET.Build.Containers.Filesystem/TestInitialization.cs
@@ -6,12 +6,12 @@ public static class TestInitialization {
     [AssemblyInitialize]
     public static void Initialize(TestContext ctx) {
         DockerRegistryManager.StartAndPopulateDockerRegistry(ctx);
-        Evaluator.LocateMSBuild(ctx);
+        ProjectInitializer.LocateMSBuild(ctx);
     }
 
     [AssemblyCleanup]
     public static void Cleanup() {
         DockerRegistryManager.ShutdownDockerRegistry();
-        Evaluator.Cleanup();
+        ProjectInitializer.Cleanup();
     }
 }
